move consensus methods to Reactor, remove conn from ConsensusState
