+++ core/indexer/src/reactor/mod.rs
@@ -8,6 +8,9 @@ pub mod mock_bitcoin;
 mod reactor_cluster_tests;
 pub mod types;
 
+use std::collections::HashSet;
+use std::time::Instant;
+
 use anyhow::{Context, Result, bail};
 use futures_util::future::pending;
 use indexer_types::{Block, BlockRow, Event, OpWithResult};
@@ -23,11 +26,19 @@ use tokio_util::sync::CancellationToken;
 
 use bitcoin::hashes::Hash;
 use bitcoin::{BlockHash, Txid};
+use malachitebft_app_channel::app::types::codec::Codec;
 use malachitebft_app_channel::app::types::core::VotingPower;
-use malachitebft_core_types::LinearTimeouts;
+use malachitebft_app_channel::app::types::sync::RawDecidedValue;
+use malachitebft_app_channel::app::types::{LocallyProposedValue, ProposedValue};
+use malachitebft_app_channel::{AppMsg, NetworkMsg};
+use malachitebft_core_types::{LinearTimeouts, Round, Validity};
+use malachitebft_engine::host::Next;
+use prost::Message;
 use tracing::{debug, error, info, warn};
 
-use crate::consensus::finality_types::StateEvent;
+use crate::consensus::codec::{ProtobufCodec, encode_commit_certificate};
+use crate::consensus::finality_types::{FINALITY_WINDOW, StateEvent, UnfinalizedBatch};
+use crate::consensus::{BatchTx, Value};
 use crate::{
     bitcoin_follower::event::{BlockEvent, MempoolEvent},
     consensus::{Genesis, Validator, ValidatorSet, signing::PublicKey},
@@ -35,7 +46,8 @@ use crate::{
         self,
         queries::{
             confirm_transaction, get_transaction_by_txid, insert_batch, insert_block,
-            insert_transaction, rollback_to_height, select_block_at_height, select_block_latest,
+            insert_transaction, insert_unconfirmed_batch_tx, rollback_to_height,
+            select_block_at_height, select_block_latest, select_existing_txids,
             select_unconfirmed_batch_tx,
         },
     },
@@ -370,7 +382,7 @@ impl<E: Executor> Reactor<E> {
             self.consensus.current_validator_set = vs;
         }
 
-        let checkpoint = self.consensus.get_checkpoint().await;
+        let checkpoint = self.consensus.get_checkpoint(&self.db_conn()).await;
         self.consensus.emit_state_event(StateEvent::BlockProcessed {
             height,
             unbatched_count,
@@ -403,6 +415,841 @@ impl<E: Executor> Reactor<E> {
         );
 
         Ok(())
+    }
+
+    async fn process_decided_batch(
+        &mut self,
+        anchor_height: u64,
+        anchor_hash: BlockHash,
+        consensus_height: crate::consensus::Height,
+        certificate: &[u8],
+        batch_txs: &[bitcoin::Transaction],
+    ) -> Result<()> {
+        let conn = self.db_conn();
+
+        info!(
+            %consensus_height,
+            anchor_height,
+            %anchor_hash,
+            num_txs = batch_txs.len(),
+            "Processing decided batch"
+        );
+
+        // Empty batch — just record for sync, no execution or finality tracking
+        if batch_txs.is_empty() {
+            insert_batch(
+                &conn,
+                consensus_height.as_u64() as i64,
+                anchor_height as i64,
+                &anchor_hash.to_string(),
+                certificate,
+                false,
+            )
+            .await
+            .context("Failed to insert empty batch")?;
+
+            let checkpoint = self.consensus.get_checkpoint(&conn).await;
+            self.consensus.emit_state_event(StateEvent::BatchApplied {
+                consensus_height,
+                anchor_height,
+                txid_count: 0,
+                checkpoint,
+            });
+
+            info!(
+                anchor = anchor_height,
+                consensus_height = %consensus_height,
+                "Empty batch recorded"
+            );
+            return Ok(());
+        }
+
+        let parsed_txs: Vec<indexer_types::Transaction> = batch_txs
+            .iter()
+            .filter_map(|btx| self.executor.parse_transaction(btx))
+            .collect();
+
+        // Track for finality — must happen once per batch execution
+        let txids: Vec<Txid> = batch_txs.iter().map(|tx| tx.compute_txid()).collect();
+        self.consensus.unfinalized_batches.push(UnfinalizedBatch {
+            consensus_height,
+            anchor_height,
+            anchor_hash,
+            txids,
+            deadline: anchor_height + FINALITY_WINDOW,
+        });
+
+        self.runtime
+            .storage
+            .savepoint()
+            .await
+            .context("Failed to begin batch transaction")?;
+
+        insert_batch(
+            &conn,
+            consensus_height.as_u64() as i64,
+            anchor_height as i64,
+            &anchor_hash.to_string(),
+            certificate,
+            false,
+        )
+        .await
+        .context("Failed to insert batch")?;
+
+        // Store raw txs for replay/sync recovery
+        for raw_tx in batch_txs {
+            let txid = raw_tx.compute_txid();
+            let serialized = bitcoin::consensus::serialize(raw_tx);
+            insert_unconfirmed_batch_tx(
+                &conn,
+                &txid.to_string(),
+                consensus_height.as_u64() as i64,
+                &serialized,
+            )
+            .await
+            .context("Failed to insert unconfirmed batch tx")?;
+        }
+
+        for t in &parsed_txs {
+            let tx_id = insert_transaction(
+                &conn,
+                indexer_types::TransactionRow::builder()
+                    .height(anchor_height as i64)
+                    .batch_height(consensus_height.as_u64() as i64)
+                    .txid(t.txid.to_string())
+                    .build(),
+            )
+            .await
+            .context("Failed to insert transaction")?;
+
+            self.executor
+                .execute_transaction(&mut self.runtime, anchor_height as i64, tx_id, t)
+                .await;
+        }
+
+        self.runtime
+            .storage
+            .commit()
+            .await
+            .context("Failed to commit batch transaction")?;
+
+        let checkpoint = self.consensus.get_checkpoint(&conn).await;
+        self.consensus.emit_state_event(StateEvent::BatchApplied {
+            consensus_height,
+            anchor_height,
+            txid_count: parsed_txs.len(),
+            checkpoint,
+        });
+
+        info!(
+            anchor = anchor_height,
+            consensus_height = %consensus_height,
+            "Batch processing complete"
+        );
+
+        Ok(())
+    }
+
+    async fn make_value(&mut self) -> Option<crate::consensus::Value> {
+        let conn = self.db_conn();
+        let last_height = self.last_height;
+        let last_hash = self.last_hash.unwrap_or(BlockHash::all_zeros());
+
+        // If blocks are pending, always propose the next one first
+        if let Some((&height, block)) = self.consensus.pending_blocks.first_key_value() {
+            return Some(crate::consensus::Value::new_block(height, block.hash));
+        }
+
+        // Pre-filter already-processed txids to avoid unnecessary validation
+        let pending_txids: Vec<Txid> = self
+            .consensus
+            .pending_transactions
+            .keys()
+            .copied()
+            .collect();
+        let txid_strs: Vec<String> = pending_txids.iter().map(|t| t.to_string()).collect();
+        let existing = select_existing_txids(&conn, &txid_strs)
+            .await
+            .unwrap_or_default();
+        let unbatched_set: HashSet<Txid> = pending_txids
+            .into_iter()
+            .filter(|t| !existing.contains(&t.to_string()))
+            .collect();
+
+        // Remove already-processed txids from the pool
+        for txid_str in &existing {
+            if let Ok(txid) = txid_str.parse::<Txid>() {
+                self.consensus.pending_transactions.remove(&txid);
+            }
+        }
+
+        // Per-tx validation — remove invalid txs from the pool
+        let mut txs = Vec::new();
+        let mut invalid_txids = Vec::new();
+        for (raw_tx, parsed) in self.consensus.pending_transactions.values() {
+            let txid = raw_tx.compute_txid();
+            if !unbatched_set.contains(&txid) {
+                continue;
+            }
+            if self.executor.validate_transaction(raw_tx, parsed).await {
+                txs.push(raw_tx.clone());
+            } else {
+                invalid_txids.push(txid);
+            }
+        }
+        for txid in &invalid_txids {
+            self.consensus.pending_transactions.remove(txid);
+        }
+        if txs.is_empty() {
+            return None;
+        }
+
+        // Batch-level validation (already-processed check will pass since we pre-filtered)
+        let candidate_txids: Vec<String> =
+            txs.iter().map(|tx| tx.compute_txid().to_string()).collect();
+        if let Some(reason) = self
+            .consensus
+            .validate_batch(
+                &conn,
+                last_height,
+                last_hash,
+                &candidate_txids,
+                last_height,
+                last_hash,
+            )
+            .await
+        {
+            info!("Not proposing batch: {reason}");
+            return None;
+        }
+
+        let value = crate::consensus::Value::new_batch_raw(last_height, last_hash, txs);
+        Some(value)
+    }
+
+    async fn validate_and_accept_proposal(
+        &mut self,
+        data: &crate::consensus::ProposalData,
+        height: crate::consensus::Height,
+        round: malachitebft_core_types::Round,
+    ) -> Option<ProposedValue<crate::consensus::Ctx>> {
+        let conn = self.db_conn();
+        let last_height = self.last_height;
+        let last_hash = self.last_hash.unwrap_or(BlockHash::all_zeros());
+
+        let value = match data {
+            crate::consensus::ProposalData::Block { height: bh, hash } => {
+                if let Some(block) = self.consensus.pending_blocks.get(bh) {
+                    if block.hash != *hash {
+                        warn!(
+                            block_height = bh,
+                            proposed = %hash,
+                            local = %block.hash,
+                            "Rejecting block proposal: hash mismatch"
+                        );
+                        return None;
+                    }
+                } else {
+                    warn!(
+                        block_height = bh,
+                        "Rejecting block proposal: block not yet received"
+                    );
+                    return None;
+                }
+                crate::consensus::Value::new_block(*bh, *hash)
+            }
+            crate::consensus::ProposalData::Batch {
+                anchor_height,
+                anchor_hash,
+                transactions,
+            } => {
+                let txid_strs: Vec<String> = transactions
+                    .iter()
+                    .map(|tx| tx.compute_txid().to_string())
+                    .collect();
+                if let Some(reason) = self
+                    .consensus
+                    .validate_batch(
+                        &conn,
+                        *anchor_height,
+                        *anchor_hash,
+                        &txid_strs,
+                        last_height,
+                        last_hash,
+                    )
+                    .await
+                {
+                    warn!("Rejecting batch proposal: {reason}");
+                    return None;
+                }
+                for tx in transactions {
+                    let txid = tx.compute_txid();
+                    let parsed =
+                        if let Some((_, cached)) = self.consensus.pending_transactions.get(&txid) {
+                            cached.clone()
+                        } else if let Some(p) = self.executor.parse_transaction(tx) {
+                            p
+                        } else {
+                            warn!(%txid, "Rejecting proposal: transaction failed to parse");
+                            return None;
+                        };
+                    if !self.executor.validate_transaction(tx, &parsed).await {
+                        warn!(%txid, "Rejecting proposal: transaction failed validation");
+                        return None;
+                    }
+                    self.consensus
+                        .pending_transactions
+                        .entry(txid)
+                        .or_insert_with(|| (tx.clone(), parsed));
+                }
+                crate::consensus::Value::new_batch_raw(
+                    *anchor_height,
+                    *anchor_hash,
+                    transactions.clone(),
+                )
+            }
+        };
+
+        let proposed = ProposedValue {
+            height,
+            round,
+            valid_round: malachitebft_core_types::Round::Nil,
+            proposer: self.consensus.address,
+            value,
+            validity: Validity::Valid,
+        };
+        self.consensus
+            .undecided
+            .insert((height, round), proposed.clone());
+        Some(proposed)
+    }
+
+    async fn try_fulfill_pending_proposal(&mut self) -> Result<bool> {
+        let last_height = self.last_height;
+        let last_hash = self.last_hash.unwrap_or(BlockHash::all_zeros());
+
+        let (past_deadline, pending_height, pending_round) = match &self.consensus.pending_proposal
+        {
+            Some(p) => (Instant::now() >= p.hard_deadline(), p.height, p.round),
+            None => return Ok(false),
+        };
+
+        let value = if let Some(value) = self.make_value().await {
+            value
+        } else if past_deadline {
+            info!(
+                height = %pending_height,
+                round = %pending_round,
+                "Proposing empty batch at hard deadline"
+            );
+            crate::consensus::Value::new_batch_raw(last_height, last_hash, vec![])
+        } else {
+            return Ok(false);
+        };
+
+        let pending = self.consensus.pending_proposal.take().unwrap();
+        let proposed = ProposedValue {
+            height: pending.height,
+            round: pending.round,
+            valid_round: Round::Nil,
+            proposer: self.consensus.address,
+            value: value.clone(),
+            validity: Validity::Valid,
+        };
+        self.consensus
+            .undecided
+            .insert((pending.height, pending.round), proposed);
+        let proposal = LocallyProposedValue::new(pending.height, pending.round, value);
+        for stream_msg in self.consensus.stream_proposal(&proposal, Round::Nil) {
+            self.consensus
+                .channels
+                .network
+                .send(malachitebft_app_channel::NetworkMsg::PublishProposalPart(
+                    stream_msg,
+                ))
+                .await
+                .context("Failed to send proposal part to network")?;
+        }
+        let _ = pending.reply.send(proposal);
+
+        Ok(true)
+    }
+
+    async fn initiate_rollback(
+        &mut self,
+        from_anchor: u64,
+        excluded_txids: HashSet<Txid>,
+    ) -> Result<()> {
+        let conn = self.db_conn();
+        let replay_batches = self
+            .consensus
+            .get_decided_from_anchor(&conn, from_anchor)
+            .await
+            .context("Failed to load replay batches for rollback")?;
+
+        info!(
+            from_anchor,
+            replay_batches = replay_batches.len(),
+            excluded = excluded_txids.len(),
+            "Initiating rollback"
+        );
+
+        let mut deferred: std::collections::VecDeque<consensus::DeferredDecision> =
+            replay_batches.into();
+        if !excluded_txids.is_empty() {
+            for decision in &mut deferred {
+                if let crate::consensus::Value::Batch { ref mut txs, .. } = decision.value {
+                    txs.retain(|tx| !excluded_txids.contains(&tx.txid()));
+                }
+            }
+        }
+        self.consensus.deferred_decisions = deferred;
+        self.consensus
+            .unfinalized_batches
+            .retain(|b| b.anchor_height < from_anchor);
+
+        self.executor
+            .replay_blocks_from(from_anchor)
+            .await
+            .context("Failed to send replay request")?;
+        Ok(())
+    }
+
+    async fn handle_consensus_msg(
+        &mut self,
+        msg: AppMsg<crate::consensus::Ctx>,
+    ) -> Result<consensus::ConsensusResult> {
+        let conn = self.db_conn();
+        let last_height = self.last_height;
+
+        let mut result = consensus::ConsensusResult::None;
+        match msg {
+            AppMsg::ConsensusReady { reply } => {
+                let start_height = self.consensus.current_height;
+                info!(%start_height, "Consensus is ready");
+
+                reply
+                    .send((start_height, self.consensus.height_params()))
+                    .map_err(|_| anyhow::anyhow!("Failed to send ConsensusReady reply"))?;
+            }
+
+            AppMsg::StartedRound {
+                height,
+                round,
+                proposer,
+                role,
+                reply_value,
+            } => {
+                info!(%height, %round, %proposer, ?role, "Started round");
+                self.consensus.current_height = height;
+                self.consensus.current_round = round;
+
+                if let Some(pending) = &self.consensus.pending_proposal
+                    && (pending.height != height || pending.round != round)
+                {
+                    info!(
+                        pending_height = %pending.height,
+                        pending_round = %pending.round,
+                        "Clearing stale pending proposal"
+                    );
+                    self.consensus.pending_proposal = None;
+                }
+
+                let proposals: Vec<_> = self
+                    .consensus
+                    .undecided
+                    .get(&(height, round))
+                    .cloned()
+                    .into_iter()
+                    .collect();
+
+                reply_value
+                    .send(proposals)
+                    .map_err(|_| anyhow::anyhow!("Failed to send StartedRound reply"))?;
+            }
+
+            AppMsg::GetValue {
+                height,
+                round,
+                timeout,
+                reply,
+            } => {
+                info!(%height, %round, "Building value to propose");
+
+                if let Some(existing) = self.consensus.undecided.get(&(height, round)) {
+                    let proposal = LocallyProposedValue::new(
+                        existing.height,
+                        existing.round,
+                        existing.value.clone(),
+                    );
+                    for stream_msg in self.consensus.stream_proposal(&proposal, Round::Nil) {
+                        self.consensus
+                            .channels
+                            .network
+                            .send(NetworkMsg::PublishProposalPart(stream_msg))
+                            .await
+                            .context("Failed to send proposal part to network")?;
+                    }
+                    reply
+                        .send(proposal)
+                        .map_err(|_| anyhow::anyhow!("Failed to send GetValue reply"))?;
+                } else if let Some(value) = self.make_value().await {
+                    let proposed = ProposedValue {
+                        height,
+                        round,
+                        valid_round: Round::Nil,
+                        proposer: self.consensus.address,
+                        value: value.clone(),
+                        validity: Validity::Valid,
+                    };
+                    self.consensus.undecided.insert((height, round), proposed);
+                    let proposal = LocallyProposedValue::new(height, round, value);
+                    for stream_msg in self.consensus.stream_proposal(&proposal, Round::Nil) {
+                        self.consensus
+                            .channels
+                            .network
+                            .send(NetworkMsg::PublishProposalPart(stream_msg))
+                            .await
+                            .context("Failed to send proposal part to network")?;
+                    }
+                    reply
+                        .send(proposal)
+                        .map_err(|_| anyhow::anyhow!("Failed to send GetValue reply"))?;
+                } else {
+                    info!(%height, %round, "Nothing to propose, holding reply for pending transactions");
+                    self.consensus.pending_proposal = Some(consensus::PendingProposal {
+                        height,
+                        round,
+                        reply,
+                        timeout,
+                        created_at: Instant::now(),
+                    });
+                }
+            }
+
+            AppMsg::ReceivedProposalPart {
+                from: _,
+                part,
+                reply,
+            } => {
+                let height = self.consensus.current_height;
+                let round = self.consensus.current_round;
+
+                let proposed = if round == Round::Nil {
+                    None
+                } else {
+                    match &part.content {
+                        malachitebft_app_channel::app::streaming::StreamContent::Data(
+                            crate::consensus::ProposalPart::Data(data),
+                        ) => {
+                            if !self.consensus.undecided.contains_key(&(height, round)) {
+                                self.validate_and_accept_proposal(data, height, round).await
+                            } else {
+                                None
+                            }
+                        }
+                        _ => None,
+                    }
+                };
+
+                reply
+                    .send(proposed)
+                    .map_err(|_| anyhow::anyhow!("Failed to send ReceivedProposalPart reply"))?;
+            }
+
+            AppMsg::ExtendVote { reply, .. } => {
+                reply
+                    .send(None)
+                    .map_err(|_| anyhow::anyhow!("Failed to send ExtendVote reply"))?;
+            }
+
+            AppMsg::VerifyVoteExtension { reply, .. } => {
+                reply
+                    .send(Ok(()))
+                    .map_err(|_| anyhow::anyhow!("Failed to send VerifyVoteExtension reply"))?;
+            }
+
+            AppMsg::Decided {
+                certificate,
+                extensions: _,
+            } => {
+                info!(
+                    height = %certificate.height,
+                    round = %certificate.round,
+                    value = %certificate.value_id,
+                    "Decided"
+                );
+            }
+
+            AppMsg::Finalized {
+                certificate,
+                extensions: _,
+                evidence,
+                reply,
+            } => {
+                info!(
+                    height = %certificate.height,
+                    round = %certificate.round,
+                    value = %certificate.value_id,
+                    evidence = ?evidence,
+                    "Finalized"
+                );
+
+                if let Some(proposal) = self
+                    .consensus
+                    .undecided
+                    .remove(&(certificate.height, certificate.round))
+                {
+                    if let Some(obs) = &self.consensus.observation {
+                        let _ = obs.decided_tx.try_send(
+                            crate::consensus::finality_types::DecidedBatch {
+                                validator_index: self.consensus.validator_index,
+                                consensus_height: certificate.height,
+                                value: proposal.value.clone(),
+                            },
+                        );
+                    }
+                    match &proposal.value {
+                        Value::Batch {
+                            anchor_height,
+                            anchor_hash,
+                            txs,
+                        } => {
+                            let mut full_txs = Vec::new();
+                            for entry in txs {
+                                match entry {
+                                    BatchTx::Raw(tx) => full_txs.push(tx.clone()),
+                                    BatchTx::Id(txid) => {
+                                        if let Some(tx) =
+                                            self.executor.resolve_transaction(txid).await
+                                        {
+                                            full_txs.push(tx);
+                                        }
+                                    }
+                                }
+                            }
+
+                            for tx in &full_txs {
+                                self.consensus
+                                    .pending_transactions
+                                    .remove(&tx.compute_txid());
+                            }
+
+                            let cert_bytes = encode_commit_certificate(&certificate)
+                                .map(|p| p.encode_to_vec())
+                                .unwrap_or_default();
+
+                            if *anchor_height < last_height {
+                                warn!(
+                                    anchor = anchor_height,
+                                    last_height,
+                                    consensus_height = %certificate.height,
+                                    "Skipping stale batch — anchor below current height"
+                                );
+                                insert_batch(
+                                    &conn,
+                                    certificate.height.as_u64() as i64,
+                                    *anchor_height as i64,
+                                    &anchor_hash.to_string(),
+                                    &cert_bytes,
+                                    false,
+                                )
+                                .await
+                                .context("Failed to record stale batch for sync")?;
+                            } else if *anchor_height > last_height {
+                                info!(
+                                    anchor = anchor_height,
+                                    last_height,
+                                    consensus_height = %certificate.height,
+                                    "Deferring batch — anchor block not yet processed"
+                                );
+                                self.consensus.deferred_decisions.push_back(
+                                    consensus::DeferredDecision {
+                                        consensus_height: certificate.height,
+                                        value: proposal.value.clone(),
+                                        certificate: cert_bytes,
+                                    },
+                                );
+                            } else {
+                                self.process_decided_batch(
+                                    *anchor_height,
+                                    *anchor_hash,
+                                    certificate.height,
+                                    &cert_bytes,
+                                    &full_txs,
+                                )
+                                .await
+                                .context("process_decided_batch failed in Finalized handler")?;
+                                result = consensus::ConsensusResult::BatchProcessed {
+                                    txids: full_txs
+                                        .iter()
+                                        .map(|tx| tx.compute_txid().to_string())
+                                        .collect(),
+                                };
+                            }
+                        }
+                        Value::Block { height, hash } => {
+                            let cert_bytes = encode_commit_certificate(&certificate)
+                                .map(|p| p.encode_to_vec())
+                                .unwrap_or_default();
+
+                            if let Some(block) = self.consensus.pending_blocks.remove(height) {
+                                info!(
+                                    block_height = height,
+                                    block_hash = %hash,
+                                    consensus_height = %certificate.height,
+                                    "Block decided and ready to process"
+                                );
+                                result = consensus::ConsensusResult::Block(
+                                    block,
+                                    consensus::DeferredDecision {
+                                        consensus_height: certificate.height,
+                                        value: proposal.value.clone(),
+                                        certificate: cert_bytes.clone(),
+                                    },
+                                );
+                            } else {
+                                let is_stale =
+                                    match select_block_at_height(&conn, *height as i64).await {
+                                        Ok(Some(row)) => row.hash != *hash,
+                                        _ => false,
+                                    };
+                                if !is_stale {
+                                    info!(
+                                        block_height = height,
+                                        block_hash = %hash,
+                                        consensus_height = %certificate.height,
+                                        "Block decided but not yet received — deferring"
+                                    );
+                                    self.consensus.deferred_decisions.push_back(
+                                        consensus::DeferredDecision {
+                                            consensus_height: certificate.height,
+                                            value: proposal.value.clone(),
+                                            certificate: cert_bytes.clone(),
+                                        },
+                                    );
+                                } else {
+                                    warn!(
+                                        block_height = height,
+                                        block_hash = %hash,
+                                        "Ignoring stale block decision (post-rollback)"
+                                    );
+                                }
+                            }
+                        }
+                    }
+                }
+
+                self.consensus.current_height = certificate.height.increment();
+                self.consensus.current_round = Round::Nil;
+                self.consensus.pending_proposal = None;
+
+                let next = Next::Start(
+                    self.consensus.current_height,
+                    self.consensus.height_params(),
+                );
+
+                reply
+                    .send(next)
+                    .map_err(|_| anyhow::anyhow!("Failed to send Finalized reply"))?;
+            }
+
+            AppMsg::GetHistoryMinHeight { reply } => {
+                let min = self
+                    .consensus
+                    .min_decided_height(&conn)
+                    .await
+                    .unwrap_or(crate::consensus::Height::new(1));
+                reply
+                    .send(min)
+                    .map_err(|_| anyhow::anyhow!("Failed to send GetHistoryMinHeight reply"))?;
+            }
+
+            AppMsg::GetDecidedValues { range, reply } => {
+                let decided = self
+                    .consensus
+                    .get_decided_range(&conn, *range.start(), *range.end())
+                    .await;
+                let values: Vec<_> = decided
+                    .into_iter()
+                    .filter_map(|(value, cert)| {
+                        ProtobufCodec
+                            .encode(&value)
+                            .ok()
+                            .map(|encoded| RawDecidedValue {
+                                certificate: cert,
+                                value_bytes: encoded,
+                            })
+                    })
+                    .collect();
+                reply
+                    .send(values)
+                    .map_err(|_| anyhow::anyhow!("Failed to send GetDecidedValues reply"))?;
+            }
+
+            AppMsg::ProcessSyncedValue {
+                height,
+                round,
+                proposer,
+                value_bytes,
+                reply,
+            } => {
+                let result: Option<ProposedValue<crate::consensus::Ctx>> =
+                    if let Ok(value) = ProtobufCodec.decode(value_bytes) {
+                        let proposed = ProposedValue {
+                            height,
+                            round,
+                            valid_round: Round::Nil,
+                            proposer,
+                            value,
+                            validity: Validity::Valid,
+                        };
+                        self.consensus
+                            .undecided
+                            .insert((height, round), proposed.clone());
+                        Some(proposed)
+                    } else {
+                        None
+                    };
+
+                reply
+                    .send(result)
+                    .map_err(|_| anyhow::anyhow!("Failed to send ProcessSyncedValue reply"))?;
+            }
+
+            AppMsg::RestreamProposal {
+                height,
+                round,
+                valid_round,
+                address: _,
+                value_id,
+            } => {
+                let lookup_round = if valid_round == Round::Nil {
+                    round
+                } else {
+                    valid_round
+                };
+                if let Some(proposal) = self.consensus.undecided.get(&(height, lookup_round))
+                    && proposal.value.id() == value_id
+                {
+                    let locally_proposed =
+                        LocallyProposedValue::new(height, round, proposal.value.clone());
+                    for stream_msg in self
+                        .consensus
+                        .stream_proposal(&locally_proposed, valid_round)
+                    {
+                        self.consensus
+                            .channels
+                            .network
+                            .send(NetworkMsg::PublishProposalPart(stream_msg))
+                            .await
+                            .context("Failed to send proposal part to network")?;
+                    }
+                }
+            }
+        }
+
+        Ok(result)
     }
 
     async fn drain_deferred_decisions(&mut self) -> Result<()> {
@@ -476,10 +1323,7 @@ impl<E: Executor> Reactor<E> {
                                 }
                             }
                         }
-                        let cs = &mut self.consensus;
-                        cs.process_decided_batch(
-                            &self.executor,
-                            &mut self.runtime,
+                        self.process_decided_batch(
                             anchor_height,
                             anchor_hash,
                             decision.consensus_height,
@@ -540,12 +1384,7 @@ impl<E: Executor> Reactor<E> {
                     .context("drain_deferred_decisions failed after block insert")?;
                 // A pending block may be what we're waiting to propose
                 if self.consensus.pending_proposal.is_some() {
-                    self.consensus
-                        .try_fulfill_pending_proposal(
-                            &self.executor,
-                            self.last_height,
-                            self.last_hash.unwrap_or(BlockHash::all_zeros()),
-                        )
+                    self.try_fulfill_pending_proposal()
                         .await
                         .context("try_fulfill_pending_proposal failed after block insert")?;
                 }
@@ -556,7 +1395,7 @@ impl<E: Executor> Reactor<E> {
                     .await
                     .context("rollback failed during Bitcoin reorg")?;
                 self.consensus.clear_on_rollback();
-                let checkpoint = self.consensus.get_checkpoint().await;
+                let checkpoint = self.consensus.get_checkpoint(&self.db_conn()).await;
                 self.consensus
                     .emit_state_event(StateEvent::RollbackExecuted {
                         to_anchor: to_height,
@@ -584,7 +1423,7 @@ impl<E: Executor> Reactor<E> {
 
             let hard_deadline_instant = self.consensus.pending_proposal.as_ref().map(|p| {
                 let deadline = p.hard_deadline();
-                let remaining = deadline.saturating_duration_since(std::time::Instant::now());
+                let remaining = deadline.saturating_duration_since(Instant::now());
                 tokio::time::Instant::now() + remaining
             });
 
@@ -654,32 +1493,16 @@ impl<E: Executor> Reactor<E> {
                 }
                 _ = debounce_sleep => {
                     debounce_deadline = None;
-                    self.consensus.try_fulfill_pending_proposal(
-                        &self.executor,
-                        self.last_height,
-                        self.last_hash.unwrap_or(BlockHash::all_zeros()),
-                    ).await
-                    .context("try_fulfill_pending_proposal failed (debounce)")?;
+                    self.try_fulfill_pending_proposal().await
+                        .context("try_fulfill_pending_proposal failed (debounce)")?;
                 }
                 _ = hard_deadline_sleep => {
-                    self.consensus.try_fulfill_pending_proposal(
-                        &self.executor,
-                        self.last_height,
-                        self.last_hash.unwrap_or(BlockHash::all_zeros()),
-                    ).await
-                    .context("try_fulfill_pending_proposal failed (hard deadline)")?;
+                    self.try_fulfill_pending_proposal().await
+                        .context("try_fulfill_pending_proposal failed (hard deadline)")?;
                 }
                 Some(msg) = consensus_rx => {
                     debug!("REACTOR: processing consensus msg");
-                    let cs = &mut self.consensus;
-                    let consensus_result = cs
-                        .handle_consensus_msg(
-                            &self.executor,
-                            &mut self.runtime,
-                            msg,
-                            self.last_height,
-                            self.last_hash.unwrap_or(BlockHash::all_zeros()),
-                        )
+                    let consensus_result = self.handle_consensus_msg(msg)
                         .await
                         .context("handle_consensus_msg failed")?;
                     if let consensus::ConsensusResult::BatchProcessed { txids } = &consensus_result
@@ -701,16 +1524,14 @@ impl<E: Executor> Reactor<E> {
                             .await
                             .context("drain_deferred_decisions failed after consensus block")?;
                         // Check finality after block execution — batch txids may now be confirmed
-                        let cs = &mut self.consensus;
-                        if cs
+                        let conn = self.db_conn();
+                        if self.consensus
                             .unfinalized_batches
                             .iter()
                             .any(|b| b.deadline <= self.last_height)
-                            && let Some((rollback_anchor, excluded)) = cs.run_finality_checks(self.last_height).await {
+                            && let Some((rollback_anchor, excluded)) = self.consensus.run_finality_checks(&conn, self.last_height).await {
                                 // Read replay decisions from DB before deleting state
-                                let cs = &mut self.consensus;
-                                cs.initiate_rollback(
-                                    &mut self.executor,
+                                self.initiate_rollback(
                                     rollback_anchor,
                                     excluded,
                                 ).await
@@ -723,9 +1544,8 @@ impl<E: Executor> Reactor<E> {
                                     .context("rollback failed during finality rollback")?;
 
                                 // Emit rollback event
-                                let cs = &mut self.consensus;
-                                let checkpoint = cs.get_checkpoint().await;
-                                cs.emit_state_event(StateEvent::RollbackExecuted {
+                                let checkpoint = self.consensus.get_checkpoint(&conn).await;
+                                self.consensus.emit_state_event(StateEvent::RollbackExecuted {
                                     to_anchor: rollback_anchor,
                                     entries_removed: 0,
                                     checkpoint,
